+++ dotnet/Services/FedExRateRequest.cs
@@ -47,8 +47,11 @@
                 //Console.WriteLine(jsonrequest);
                 Stopwatch stopWatch = new Stopwatch();
                 stopWatch.Start();
+                // Console.WriteLine(JsonConvert.SerializeObject(getRatesRequest));
+                // Console.WriteLine(JsonConvert.SerializeObject(request));
                 getRatesResponse ratesResponse = await client.getRatesAsync(request);
                 stopWatch.Stop();
+                // Console.Write(JsonConvert.SerializeObject(ratesResponse));
                 // Get the elapsed time as a TimeSpan value.
                 TimeSpan ts = stopWatch.Elapsed;
                 getRatesResponseWrapper.timeSpan = ts;
@@ -60,33 +63,38 @@
                 {
                     ShowRateReply(reply);
                     //string replyjson = JsonConvert.SerializeObject(reply);
+                    int totalQuantity = 0;
+                    foreach (Item item in getRatesRequest.items) {
+                        totalQuantity += item.quantity;
+                    }
                     foreach(RateReplyDetail detail in reply.RateReplyDetails)
                     {
                         //Item matchingItem = getRatesRequest.items.Select(x => x.id.Equals())
-                        //GetRatesResponse rateResponse = new GetRatesResponse
                         //{
-                        //    carrierId = detail.ServiceType,
-                        //    carrierName = CARRIER,
-                        //    deliveryOnWeekends = false,
-                        //    dockId = getRatesRequest.items[0].availability[0].dockId,
-                        //    itemId = getRatesRequest.items[0].id,
-                        //    price = detail.RatedShipmentDetails[0].ShipmentRateDetail.TotalNetCharge.Amount,
                         //    quantity = getRatesRequest.items[0].quantity,
                         //    time = $"{this.TransitDays(detail.ServiceType, detail.TransitTime.ToString(), detail.DeliveryTimestamp.ToString())}.00:00:00"
                         //};
-
+                        TimeSpan transitArrival = detail.DeliveryTimestamp - getRatesRequest.shippingDateUTC;
                         GetRatesResponse rateResponse = new GetRatesResponse
                         {
-                            carrierId = detail.ServiceDescription.Code,
-                            carrierName = detail.ServiceDescription.Description,
-                            //deliveryOnWeekends = false,
+                            carrierId = "FEDEX",
                             //dockId = getRatesRequest.items[0].availability[0].dockId,
-                            //itemId = getRatesRequest.items[0].id,
+                            itemId = getRatesRequest.items[0].id,
                             price = detail.RatedShipmentDetails[0].ShipmentRateDetail.TotalNetCharge.Amount,
-                            //quantity = getRatesRequest.items[0].quantity,
-                            time = $"{this.TransitDays(detail.ServiceType, detail.TransitTime.ToString(), detail.DeliveryTimestamp.ToString())}.00:00:00",
-                            slaType = detail.ServiceType
+                            numberOfPackages = totalQuantity,
+                            estimateDate = detail.DeliveryTimestamp,
+                            shippingMethod = detail.ServiceDescription.Description,
+                            transitTime = transitArrival.ToString(),
+                            carrierSchedule = new List<Schedule>(),
+                            deliveryChannel = "delivery",
+                            weekendAndHolidays = new WeekendAndHolidays(),
+                            pickupAddress = null,
                         };
+
+                        rateResponse.carrierBusinessHours = new BusinessHour[7];
+                        for (int day = 0; day < 7; day++) {
+                            rateResponse.carrierBusinessHours[day] = new BusinessHour((DayOfWeek) day, new TimeSpan(0, 0, 0).ToString(), new TimeSpan(23, 59, 59).ToString());
+                        }
 
                         getRatesResponseWrapper.GetRatesResponses.Add(rateResponse);
                     }
@@ -152,15 +160,14 @@
                             GetRatesResponse rateResponse = new GetRatesResponse
                             {
                                 carrierId = detail.ServiceDescription.Code,
-                                carrierName = detail.ServiceDescription.Description,
                                 //deliveryOnWeekends = false,
                                 //dockId = tempRequest.items[0].availability[0].dockId,
                                 //wareHouseId = tempRequest.items[0].availability[0].warehouseId,
                                 //itemId = tempRequest.items[0].id,
                                 price = detail.RatedShipmentDetails[0].ShipmentRateDetail.TotalNetCharge.Amount,
                                 //quantity = tempRequest.items[0].quantity,
-                                time = $"{this.TransitDays(detail.ServiceType, detail.TransitTime.ToString(), detail.DeliveryTimestamp.ToString())}.00:00:00",
-                                slaType = detail.ServiceType
+                                // time = $"{this.TransitDays(detail.ServiceType, detail.TransitTime.ToString(), detail.DeliveryTimestamp.ToString())}.00:00:00",
+                                shippingMethod = detail.ServiceType
                             };
 
                             getRatesResponseWrapper.GetRatesResponses.Add(rateResponse);
@@ -212,9 +219,7 @@
             
             request.ReturnTransitAndCommit = true;
             request.ReturnTransitAndCommitSpecified = true;
-            
             SetShipmentDetails(request, getRatesRequest);
-            
             return request;
         }
 
@@ -224,8 +229,10 @@
             SetOrigin(request, getRatesRequest);
             SetDestination(request, getRatesRequest);
             SetPackageLineItems(request, getRatesRequest);
-            //request.RequestedShipment.PackageCount = getRatesRequest.items.Count.ToString();
             request.RequestedShipment.PackageCount = getRatesRequest.items.Sum(x => x.quantity).ToString();
+            request.RequestedShipment.PreferredCurrency = getRatesRequest.currency;
+            request.RequestedShipment.ShipTimestampSpecified = true;
+            request.RequestedShipment.ShipTimestamp = getRatesRequest.shippingDateUTC;
         }
 
         private void SetOrigin(RateRequest request, GetRatesRequest getRatesRequest)
@@ -237,6 +244,9 @@
             request.RequestedShipment.Shipper.Address.City = getRatesRequest.origin.city;
             request.RequestedShipment.Shipper.Address.StateOrProvinceCode = getRatesRequest.origin.state;
             request.RequestedShipment.Shipper.Address.StreetLines = new string[] { getRatesRequest.origin.street };
+            request.RequestedShipment.Shipper.Address.ResidentialSpecified = getRatesRequest.origin.residential;
+            request.RequestedShipment.Shipper.Address.Residential = getRatesRequest.origin.residential;
+            // request.RequestedShipment.Shipper.Address.GeographicCoordinates = getRatesRequest.origin.coordinates.latitude.ToString("+#.###;-#.###;0") + getRatesRequest.origin.coordinates.longitude.ToString("+#.###;-#.###;0");
         }
 
         private void SetDestination(RateRequest request, GetRatesRequest getRatesRequest)
@@ -248,6 +258,9 @@
             request.RequestedShipment.Recipient.Address.City = getRatesRequest.destination.city;
             request.RequestedShipment.Recipient.Address.StateOrProvinceCode = getRatesRequest.destination.state;
             request.RequestedShipment.Recipient.Address.StreetLines = new string[] { getRatesRequest.destination.street };
+            request.RequestedShipment.Recipient.Address.ResidentialSpecified = getRatesRequest.destination.residential;
+            request.RequestedShipment.Recipient.Address.Residential = getRatesRequest.destination.residential;
+            // request.RequestedShipment.Recipient.Address.GeographicCoordinates = getRatesRequest.destination.coordinates.latitude.ToString("+#.###;-#.###;0") + getRatesRequest.destination.coordinates.longitude.ToString("+#.###;-#.###;0");
         }
 
         private void SetPackageLineItems(RateRequest request, GetRatesRequest getRatesRequest)
@@ -264,15 +277,25 @@
                 request.RequestedShipment.RequestedPackageLineItems[cnt].Weight.Units = WeightUnits.LB;
                 request.RequestedShipment.RequestedPackageLineItems[cnt].Weight.UnitsSpecified = true;
                 //request.RequestedShipment.RequestedPackageLineItems[cnt].Weight.Value = getRatesRequest.items[cnt].unitDimension.weight * getRatesRequest.items[cnt].quantity;
-                request.RequestedShipment.RequestedPackageLineItems[cnt].Weight.Value = getRatesRequest.items[cnt].unitDimension.weight;
+                request.RequestedShipment.RequestedPackageLineItems[cnt].Weight.Value = Convert.ToDecimal(getRatesRequest.items[cnt].unitDimension.weight);
                 request.RequestedShipment.RequestedPackageLineItems[cnt].Weight.ValueSpecified = true;
                 // package dimensions
                 request.RequestedShipment.RequestedPackageLineItems[cnt].Dimensions = new Dimensions();
-                request.RequestedShipment.RequestedPackageLineItems[cnt].Dimensions.Length = getRatesRequest.items[cnt].unitDimension.length.ToString();
-                request.RequestedShipment.RequestedPackageLineItems[cnt].Dimensions.Width = getRatesRequest.items[cnt].unitDimension.weight.ToString();
-                request.RequestedShipment.RequestedPackageLineItems[cnt].Dimensions.Height = getRatesRequest.items[cnt].unitDimension.height.ToString();
+                request.RequestedShipment.RequestedPackageLineItems[cnt].Dimensions.Length = Math.Ceiling(getRatesRequest.items[cnt].unitDimension.length).ToString();
+                request.RequestedShipment.RequestedPackageLineItems[cnt].Dimensions.Width = Math.Ceiling(getRatesRequest.items[cnt].unitDimension.width).ToString();
+                request.RequestedShipment.RequestedPackageLineItems[cnt].Dimensions.Height = Math.Ceiling(getRatesRequest.items[cnt].unitDimension.height).ToString();
                 request.RequestedShipment.RequestedPackageLineItems[cnt].Dimensions.Units = LinearUnits.IN;
                 request.RequestedShipment.RequestedPackageLineItems[cnt].Dimensions.UnitsSpecified = true;
+                
+                //Special Handling goods
+                if (!String.IsNullOrEmpty(getRatesRequest.items[cnt].modal)) {
+                    request.RequestedShipment.RequestedPackageLineItems[cnt].SpecialServicesRequested = new PackageSpecialServicesRequested();
+                    request.RequestedShipment.RequestedPackageLineItems[cnt].SpecialServicesRequested.SpecialServiceTypes = new String[] {"DANGEROUS_GOODS"};
+                    request.RequestedShipment.RequestedPackageLineItems[cnt].SpecialServicesRequested.DangerousGoodsDetail = new DangerousGoodsDetail();
+                    request.RequestedShipment.RequestedPackageLineItems[cnt].SpecialServicesRequested.DangerousGoodsDetail.Offeror = "TEST OFFEROR";
+                    request.RequestedShipment.RequestedPackageLineItems[cnt].SpecialServicesRequested.DangerousGoodsDetail.EmergencyContactNumber = "3268545905";
+                    request.RequestedShipment.RequestedPackageLineItems[cnt].SpecialServicesRequested.DangerousGoodsDetail.Options = new HazardousCommodityOptionType[] { HazardousCommodityOptionType.HAZARDOUS_MATERIALS };
+                }
             }
         }