+++ CHANGELOG.md
@@ -6,3 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Modified existing request/response to accomodate dynamic rate hub request/response
+- Allowed rounding for length, width, and height
+- Modified the routing to connect with the dynamic rates hub
+
+### Fixed
+
+- Fixed a bug where Width was taking the Weight value